requirements.txt
gradio>=4.0
numpy
pillow
matplotlib
